requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic[email]
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
aiofiles==24.1.0
starlette==0.38.5
requests==2.32.3
anyio==4.6.2.post1
langfuse
# optional next milestones:
# langgraph
# pgvector
